Restructure package layout
