--- src/main/kotlin/com/skaggsm/mumblelinkmod/MumbleLink.kt
@@ -1,8 +1,0 @@
-package com.skaggsm.mumblelinkmod
-
-/**
- * Created by Mitchell Skaggs on 5/28/2019.
- */
-object MumbleLink {
-    const val MODID: String = "fabric-mumblelink-mod"
-}
