+++ src/main/kotlin/com/skaggsm/mumblelinkmod/client/ClientConfig.kt
@@ -1,0 +1,4 @@
+package com.skaggsm.mumblelinkmod.client
+
+class ClientConfig {
+}
