+++ src/main/kotlin/com/skaggsm/mumblelinkmod/client/MathUtils.kt
@@ -1,0 +1,12 @@
+package com.skaggsm.mumblelinkmod
+
+import net.minecraft.util.math.Vec3d
+
+/**
+ * Convert to a float 3-array in a left-handed coordinate system.
+ * Minecraft is right-handed by default, Mumble needs left-handed.
+ *
+ * @see <a href="https://wiki.mumble.info/wiki/Link#Coordinate_system">Coordinate system</a>
+ */
+val Vec3d.toLHArray: FloatArray
+    get() = floatArrayOf(x.toFloat(), y.toFloat(), -z.toFloat())
