+++ src/main/kotlin/com/skaggsm/mumblelinkmod/client/MumbleLinkModMenu.kt
@@ -1,7 +1,7 @@
-package com.skaggsm.mumblelinkmod.screen
+package com.skaggsm.mumblelinkmod.client
 
-import com.skaggsm.mumblelinkmod.MumbleLink
-import com.skaggsm.mumblelinkmod.config.MumbleLinkConfig
+import com.skaggsm.mumblelinkmod.main.MainMumbleLinkMod.MODID
+import com.skaggsm.mumblelinkmod.main.OldConfig
 import io.github.prospector.modmenu.api.ConfigScreenFactory
 import io.github.prospector.modmenu.api.ModMenuApi
 import me.sargunvohra.mcmods.autoconfig1u.AutoConfig
@@ -15,15 +15,15 @@ import java.util.function.Function
  */
 @Environment(EnvType.CLIENT)
 class MumbleLinkModMenu : ModMenuApi {
-    override fun getModId(): String = MumbleLink.MODID
+    override fun getModId(): String = MODID
 
     override fun getModConfigScreenFactory(): ConfigScreenFactory<*> =
             ConfigScreenFactory { screen ->
-                AutoConfig.getConfigScreen(MumbleLinkConfig::class.java, screen).get()
+                AutoConfig.getConfigScreen(OldConfig::class.java, screen).get()
             }
 
     override fun getConfigScreenFactory(): Function<Screen, out Screen> =
             Function { screen ->
-                AutoConfig.getConfigScreen(MumbleLinkConfig::class.java, screen).get()
+                AutoConfig.getConfigScreen(OldConfig::class.java, screen).get()
             }
 }
