+++ src/main/kotlin/com/skaggsm/mumblelinkmod/main/MainConfig.kt
@@ -1,0 +1,4 @@
+package com.skaggsm.mumblelinkmod.main
+
+class MainConfig {
+}
