+++ src/main/kotlin/com/skaggsm/mumblelinkmod/server/ServerConfig.kt
@@ -1,0 +1,4 @@
+package com.skaggsm.mumblelinkmod.server
+
+class ServerConfig {
+}
