+++ src/main/kotlin/com/skaggsm/mumblelinkmod/server/ServerMumbleLinkMod.kt
@@ -1,0 +1,11 @@
+package com.skaggsm.mumblelinkmod.server
+
+import net.fabricmc.api.DedicatedServerModInitializer
+
+object ServerMumbleLinkMod : DedicatedServerModInitializer {
+    /**
+     * Runs after [MainMumbleLinkMod.onInitialize].
+     */
+    override fun onInitializeServer() {
+    }
+}
